backend/requirements.txt
fastapi>=0.115.0
sqlmodel==0.0.14
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pyjwt>=2.8.0
bcrypt==4.1.2
passlib[bcrypt]>=1.7.4
